rename types dir to proto
